CI: use correct branch for testing stable branches
With this patch, we will use the correct branch of the
tests repository to test changes of other repositories,
as we cannot use master branch to test a stable branch from
another repository.

E.g. We must use 'kata-containers/tests' stable-1.1 branch
for testing changes of the 'kata-containers/runtime'
stable-1.1 branch.

Fixes: #625.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>