functional: The terminal needs to be disabled
When we use the default spec config options with terminal being
enabled, the shim errors right after it's been started. The reason
being it cannot set the terminal properly since ginkgo functional
tests never create any terminal.

With this patch, the kata-shim process runs as long as the container
workload is running, which is the expected behavior.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>