ci: cleanup: fix process kill for multiple processes
When trying to kill off stale processes with a combination of
pgrep and kill, a mix of line separator and quotations meant the
kill only worked if we found a single process.

Fix by using a separator of ' '(space) for the pgrep and
don't quote the pid expansion - a pid list is just a plain
list of numbers with spaces between them.

Fixes: #974

Signed-off-by: Graham Whaley <graham.whaley@intel.com>