integration/docker: fix update cpu tests
Update cpu tests are failing because of a race condition between the test
and the kata agent. When a new CPU is hot plugged it's not connected
automatically, kata agent must receive the order to connect it, therefore
the existence of the directory /sys/devices/system/cpu/cpuX (where X is the
number of a CPU) is not a guarantee that the cpu is online, hence the tests
shouldn't check that directory, instead the online file of the last CPU should
be inspected.

fixes #1209

Signed-off-by: Julio Montes <julio.montes@intel.com>