ci: image: get image from ci server if possible.
Today we do not generate the rootfs in the CI
if the image is not up-to-date only the agent
is open and installed. While this gives speed
to the CI it can cause issues at release where
assets are created form scratch.

Try to get image from server, if fail build from source.

Instead lets cache the image created with the new script:

.ci/ci_cache_image.sh

This is going to run in a daily CI job to reduce jobs time.

.ci/install_kata_image.sh
.ci/ci_cache_image.sh

Fixes: #1213

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>