CI: Do not assume hypervisors run with the full path
Soak tests assume hypervisors run with the full path.

However when running a hypervisor with jailer the path is relative to
the jail and hence the check should only look for the process.

For example when running firecracker with jailer the process will be

```
 /firecracker
--id=5af26fcea8dbf413067e90026fd9845515d5a1f300d61f66427760a27ef114c2
--seccomp-level=2 --start-time-us=1562786230593253
--start-time-cpu-us=223 --api-sock=/api.socket�
where /firecracker is relative to the jailed location.
```

Fixes: #1802.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>