ci: Skip vsock_test for firecracker
With the fix on issue #2383, we got a regression on the vsock_test with
CI on firecracker. The test was silently skipped because the
'DefaultHypervisor' (a.k.a 'qemu') was used as the key to read Kata
configuration even when 'firecracker' should be used as the key. With the
wrong key, the 'KataConfig.Hypervisor[KataHypervisor].Vsock' is 'false'
making the test silently skipped, while the test is not skipping and
failing with the right key. As a temporary work-around, we need to
disable the test first.

Fixes: #2397

Signed-off-by: Bo Chen <chen.bo@intel.com>