ci/openshift-ci: Onboarding into the OpenShift CI
The OpenShift CI allows for components to be built and tested in clusters
provided by the OpenShift community. This adds the minimum required on kata's side to get it onboarded in that CI system.

More precisely, it introduces the buildall_install.sh script which is used
to build and install kata-containers into a container image.

Fixes #2527

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>