ci: Consider DESTDIR to install QEMU
Currently the install_qemu and install_qemu_experimental scripts install the cached binaries if possible, and both extract the files in '/'. This was changed so that the DESTDIR variable is considered.

Changes made exclusively to install_qemu:
a) The KATA_QEMU_DESTDIR variable is used as prefix for when building the static QEMU, whereas the built for dynamic linked QEMU uses PREFIX. So the KATA_QEMU_DESTDIR was removed, instead PREFIX is used for consistency.

Fixes #2537

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>