ci/lib.sh: Make build_and_install use DESTDIR and PREFIX
The build_and_install function evokes `make` to build the component, and
the component's Makefile allows for setting the DESTDIR and PREFIX variables. However, one should pass those variables as arguments to `make` otherwise they aren't honored.

So this change the build_and_install so that DESTDIR and PREFIX are properly
passed to `make` if they are exported on current environment.

Fixes #2537

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>