ci: Make install scripts to use the DESTDIR variable
The install_runtime, install_kata_kernel and, install_kata_image all
recognize the PREFIX variable to install files. The DESTDIR is used
to configure the installation of built binaries but it hasn't been
used by those scripts. So this change allow they to place files
on DESTDIR if defined.

Fixes #2537

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>