metrics: report: report missing data files gracefully
The metrics report did not handle missing data files (that is,
partial runs and data captures) well - it generated warnings
and strange R errors in the pages of the report where there
was no relevant data.

This tended to cause general confusion, as quite often the
initial impression is that the report generation has failed
as a whole.

Improve by producing meaningful 'No data for this test' type
warnings in the final report.

Fixes: #2555

Signed-off-by: Graham Whaley <graham.whaley@intel.com>