metrics: report: fix semver updates in dut tables
Runtime commit:
https://github.com/kata-containers/runtime/commit/a4b3c65c16f63f1690966f68426a703cd3af7ab5
changed the shape of the Version info for the runtime/proxy/shim,
which then filtered out into the JSON data that the report generator
processes to geneate the 'device under test' table. It broke it.

Fix it up by adding the sub-level Semver field to extract the version as
a single whole string.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>