metrics: report: make fio percentile x-axis readable
The x-axis on the fio latency percentile graphs was using the default
divisons, which was 'divide into 8', which placed each division tickmark
at 12.5% intervals. That's not useful for visually assessing the data.
Fix the intervals at 10%, which although a touch dense and noisy, does
make it much easier to assess the graph, particularly at the 'knee'
point, which tends to be crammed into the top 10% of the data.

Fixes: #2558

Signed-off-by: Graham Whaley <graham.whaley@intel.com>