metrics: time: explicitly strip leading 0's from date
The `date` command `-N` format changed in version 8.32, and now
pre-pends 0's to pad the field. Explicitly strip those 0's, otherwise
it causes the later json prints to barf, as it sees them as octal.
This is also backwards compat with older date versions (if the 0's
are not there, the strip is benign).

Fixes: #2562

Signed-off-by: Graham Whaley <graham.whaley@intel.com>