clh: ci: Do not build image using podman
On cloud-hypervisor jobs when the image is not cached, the CI scripts try to
build the image from a source, but when it does uses PODMAN, the logic of the
script installation is that if TEST_CGROUPSV2 is false, it will use docker else
podman, the clh jobs export TEST_CGROUPSV2 as no. And fallback to use
podman.On clh jobs when the image is not cached, the CI scripts try to build
the image from a source, but when it does uses PODMAN, the logic of the script
installation is that if TEST_CGROUPSV2 is false, it will use docker else
podman, the clh jobs export TEST_CGROUPSV2 as no.

Use `false` to avoid the script use podaman.

Fixes: #2604

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>