ci: Add a entry point for CI
To modify jenkins_job_build.sh in a PR
the jobs from test repository clone test and checkout to the
repository first, this is duplicated in a lot of jobs.
Also, in CI jobs some variables that change what a job
does is hidden in the job, if we want to change how will be tested
a CI job then will need to update the variables in all the jobs.

This script should be minimal to avoid change it in the future.
So if jenkins_job_build.sh changes the changes will be taken
no matter the branch.

The script also try to simply a jenkins job internal script.

Fixes: #2652

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>