ci: VFIO: Enforce to use the host physical bits when launching L1 VM
By default, qemu will use a magic number '40' as the physical bits for
the guest vcpu even with option '-cpu host' (passthrough host
cpus). This causing problem to run clh (L2) VMs as the physical bits on
L1 can be larger than the actual supported physical bits on the
L0/bare-metal machine. This is also why clh is not runnable and the VFIO
CI is failing on the current CI machine whose host physical bits is '39'.

Signed-off-by: Bo Chen <chen.bo@intel.com>