checkcommits: fix singleWordLine
strings.TrimSpace only trims leading and trainling spaces. As a result,
singleWordLine is almost always true. Change to use strings.Fields to
determine singleWordLine.

In the meantime, change defaultMaxBodyLineLength to 150. 72 is way too
small and no one is using such small terminals. Linux kernel already
extends its 80-character limit to 100 and it is just a warning not
a real limit. Since we use it as a hard limit to filter out overly large
lines, let's use a really overly large number.

Fixes: #2847
Signed-off-by: Peng Tao <bergwolf@hyper.sh>