ci: Not set 'KATA_DEV_MODE' to 'false' explicitly
In many of our CI scripts, we assume the 'KATA_DEV_MODE' environment
variable to be not set at all when we are not in the 'kata dev
mode'. Setting its value explicitly to 'false' would fail the check '[-n
$KATA_DEV_MODE]' and make the script believe we ARE in the 'kata dev
mode', which would skip some installation. For example,
'install_rust.sh' is totally skipped in CLH's CI, which caused our
current regression.

Fixes: #2863

Signed-off-by: Bo Chen <chen.bo@intel.com>