versions: Update flannel version to v0.13.0-rc2
Recently we've seen CI issues on both Debian and CentOS because of
iptables > 1.8 has been causing a different set of connectivity issues.

The one CI is facing is due to a known issue with flannel[0], which was
pointed out by Archana Shinde and updating flannel does solve the issue
without causing any other regression.

A huge thanks to Archana for nailing down the issue.

[0]: https://github.com/coreos/flannel/issues/1306

Fixes: #2910

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>