ci: Fix cri-o drop-in configuration file
Add runc to the runtimes map in the cri-o drop-in file.
In addition, the default runtime is set to runc as we use
runc for initiating the kubernetes cluster and then we use
RuntimeClass to run tests using Kata.

Add `manage_ns_lifecycle = true` to the drop in file and remove
the commands that changed this in `.ci/install_crio.sh`

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>