ci: Fix cri-o drop-in configuration file
Add runc to the runtimes map in the cri-o drop-in file.
In addition, the default runtime is set to runc as we use
runc for initiating the kubernetes cluster and then we use
RuntimeClass to run tests using Kata.

Add `manage_ns_lifecycle = true` to the drop in file and remove
the commands that changed this in `.ci/install_crio.sh`

Depends-on: github.com/kata-containers/kata-containers#981

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>
(cherry picked from commit 0fb0f3bdd3069747091d1013bf615b5f75556294)