ci: cleanup package install for SLES
Most of the repositories that are added to the SLES nodes can be removed
as many of the additionally required packages are meanwhile part of the
"PackageHub" repositories that are offered for SLES.
This should also speed up things a bit as only a single "zypper refresh"
call is needed now.

Move packages to the "packages" list instead of separate zypper calls
where possible. Having everything in one (or two) bigger zypper calls
should make things a bit faster.

Also install yamllinit from the available SLES package instead of pypi.

Fixes #2986
Signed-off-by: Ralf Haferkamp <rhafer@suse.com>