ci/install_runtime: Allow to skip vhost setup
The install_runtime.sh script attempts to load the vhost module if USE_VSOCK is
enabled. In the OpenShift CI it can build with vsock but the host configuration
should not be changed, in fact trying to load the vhost module will end up on
failure. So this introduces the OPENSHIFT_CI variable which if set to 'true'
it will skip vhost setup in the host.

Fixes #2614

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>