ci: Use init_ci_flags for all CI_JOB cases
Let's take advantage of `init_ci_flags` function and use it for all the
CI_JOB cases.  By doing this, we also help to enforce its usage for the
next person adding a new case and looking for an example in the current
code.

Also, to help me with my OCD, I've ordered the variables in each one of
the switch cases.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>