tests/cri-containerd: Set default value to KATA_HYPERVISOR
The script uses KATA_HYPERVISOR variable to decide whether to skip or not
tests. Its value can be unset (then set to empty) and this will cause
the conditional code to not really work. To avoid that situation, this
will set a default value to KATA_HYPERVISOR, which will be "qemu" like
in the other test scripts.

Fixes #2329

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>