actions: Set up github workspace as GOPATH
Since actions/checkout2 is not checking out code under GOPATH,
go is not checking parent's vendor directory for dependency.
Because os this the checkcommits code is not able to find the required
vendored directories in the parent directory.

With this commit, we are setting the GOPATH to {{ github.workspace
}}/tests. This the path used by actions/checkout step as this action
can only download code relative to this path and not to the GOPATH
setup by previous go install.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>