ci/openshift-ci: Add the buildall_install.sh script
This introduces the buildall_install.sh script which is used
to build and install kata-containers into a well-known path
in a container image.

Forward-port of commit a37e134ef50ae5. But on kata 2.x the
buildall_install.sh doesn't install QEMU since at this point
it isn't being cached on Jenkins. It will be built in a separated
stage of the OpenShift CI pipeline.

Fixes #2527

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>