tests: Add some kubectl describe information to teardown of many k8s tests
When kubernetes tests in the CI fail, it can be difficult to work out why.
In particular, one fairly likely failure mode is for a pod to be created
but never become ready.  If this happens, the CI output will show a timeout
on the kubectl wait command, but no indication of why the pod didn't become
ready.

This adds a kubectl describe to the teardown() path for many of the tests,
which will give at least a little more indication of where to look for a
problem.  Technically the debug command will be executed for both passing
and failing tests, however bats will suppress the output for passing tests,
so it shouldn't clutter the output except in cases where it is likely to be
useful.

This is by no means attempting to add exhaustive debugging information to
all the tests, just a basic start for a good selection.  Further debug
instrumentation in the teardown paths seems likely for future improvements
to debugability.

fixes #3210

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>