test: docker: Adjust the `cpuset-cpus test`
As we are fixing the `--cpuset-cpus` support in kata with
docker (kata-containers/runtime#3138), it exposes the `cpuset-cpus test`
is actually not testing the expected behavior of kata. By design, Kata
does not constrain the guest vCPUs at all, because we don't track the
mappings between guest vCPU and host CPUs. As a result, we are always
allowing containers to use all onlined vCPUs in the guest. The current
cpuset-cpus assumes the guest cgroups constrain the vCPU in the exactly
same way as the CPUs are constrained on the host. That is wrong.

This patch adjusts the `cpuset-cpus test` to respect the actual behavior
of kata with `docker --cpuset-cpus`.

Depends-on: github.com/kata-containers/runtime#3138

Fixes: #3241

Signed-off-by: Bo Chen <chen.bo@intel.com>