ci: Apply QEMU patches on vfio jobs
The vfio_jenkins_job_build.sh is going to spawn a VM, it uses a built
from source QEMU, being the exact same version used on all the other
CI jobs. With recent bump to QEMU 5.2.0 it is needed to apply one
patch that fix the build. Thus this changed the script to apply that
and all the other patches.

Fixes #3307
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>