metrics: clh: Update the midval for memory footprint for kata 1.x
The metrics CI on cloud-hypervisor for kata 1.x is now broken because of
increased memory footprint. A similar issue on metrics CI for qemu was
also reported recently in issue #3335. Following the fix/work-around on
the qemu+kata 1.x metrics CI (PR #3336), we will need to update the mid-val
of reference memory footprint for cloud-hypervisor w/ kata 1.x. This
patch increase mid-val for cloud-hypervisor by 22.5% which is the same
increase percentage on qemu's metrics CI.

Fixes: #3379

Signed-off-by: Bo Chen <chen.bo@intel.com>