ci: Simplify CRI-O installation scripts
Some of the logic we have was introduced as part of e96858391, and
that's from the OCP 3.x era, and not actually useful for us anymore.

Also, there's a whole bunch of logic that dates CRI-O 1.10 / 1.11, which
is not even supported anymore.  Those can also be safely removed.

Let's take the advantage that we're already cleaning up the
versions.yaml file on the `kata-containers/kata-containers` side and
simplify the installation script here.

Fixes: #3412

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>