integration: Use systemd as cgroup driver for CRI-O k8s tests
Seems that cgroupfs has been used simply because systemd was not yet
stable when we introduced the `integration/kubernetes` tests.

As CRI-O uses "systemd" as the cgroup manager, let's make this
configurable and enforce CRI-O tests will be run using "systemd" as the
cgroup manager, and containerd tests will keep running using "cgroupfs".

This approach will ensure we don't regress in any of the fronts.

Depends-on: github.com/kata-containers/kata-containers#1689
Fixes: #3413

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>