integration: Remove old CRI-O tests
CRI-O integration tests are not running as part of Kata Containers 2.x
CI for some time already.  CRI-O team is working on having those tests
as part of the OpenShift CI and run those on their side.

For us, Kata Containers, we're good to go on having the integration
tests here.

Fixes: #3411

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>