kata-webhook: Use env var to change runtimeclass name
kata-webhook currently works with hardcoded RuntimeClass name kata.
If the k8s cluster is configured with other RuntimeClasses, such as
kata-clh or kata-qemu, the webhook will not work.

With this change, the webhook will be able to get the name of the
RuntimeClass from an environment variable so that you can specify
other values.

Updated documentation for this new change.

Fixes: #3415.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>