crio: Ensure crio is built with the selinux tag
Testing both kata-containers and CRI-O built with the `selinux` tag is
something we should be doing for some time already, and we've been using
this on the OpenShift side.

Mind, though, that this is a no-op for now as the OSes are set to run
SELinux on permissive mode.  This will help most folks who are running
the CI locally and checking for audit logs and whatnot has to be
improved.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>