crio: Use branch rather than a version
We're now following release-1.21 branch (according to the kubernetes
1.21 version).

The decision to use a branch rather than an official version is that we
can always get the stable fixes and do not have to do the bump on our
side.  This was discussed as part of the Architecture Committee meeting
on April 27th 2021.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>