ci/install_qemu: Use cached QEMU
This changed the installer script to fetch and install a built QEMU
if a suitable version is found on the project's Jenkins.

To detect changes that would flag a cache mismatch it is employed
a simple algorithm to generate a SHA-256 from the files and scripts
used to build QEMU. The cached and current sha-256 is compared to
determine whether or not use the cached asset.

Fixes #3119
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>