k8s: Set a minimum timeout across tests
Our e2e's use either waitForProcess() or `kubectl wait` for waiting resources on the test
cluster. Both rely on timeouts, being the later set to 30s by default and the former requires
the user to pass wait_time and sleep_time parameters. Some tests leave the defaults, others
set a value, and a bunch relies on $timeout (variable defined in .ci/lib.sh).

It isn't uncommon to see tests failing on CI because they hit a timeout. This changed all the
tests to use a mininum of 90s (current value of $timeout) on wait operatios, with the hope
those fails will become really rare.

Fixes #3496
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>