ci/install_qemu: Always copy the tarball to the cache directory
Currently when QEMU is built locally its tarball gets copied to
a cache directory on localhost so that the artifact can be archived
by the Jenkins cache job.

This changed the installer script to copy the tarball even if it
was pulled so that the Jenkins cache job still can archive it. In the end
the job will simply pull and archive (again) the same tarball, as
opposed to always build it (as the cache mechanism used to work for
kata 1.x).

Fixes #3119
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>
(cherry picked from commit d5a91e26e1fd13440aa5da4e00a7fa4f1c629147)