ci: Let CRI-O install its own systemd service file
Currently we use our own cooked version of the CRI-O's systemd service
file, and we do that for no good reason.  Instead, let's start using the
one provided by the CRI-O project.

With this, we lower the CI specific knowledge needed to run the CI
locally, either for debugging issues or simply for development / testing
purposes.

With this change, we start setting `log_level = "debug"` via the CRI-O's
configuration file, rather than hardcoding it in its systemd service
file.

Fixes: #3570

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>