ci: Enable osbuilder with Podman & w/o container
Old behaviour:
Use $osbuild_docker for $USE_DOCKER if set or set to 1. Effectively,
this means running with Docker no matter what, because osbuilder only
checks -z $USE_DOCKER.
New behaviour:
Set $USE_DOCKER to $osbuild_docker unless it is set to a falsy value or
$USE_PODMAN is already set. This allows running with
osbuild_docker={0,false,no} and running without a container engine, or
setting it to anything else and that value being passed on, or using
Podman.

Fixes: #3598

Signed-off-by: Jakob Naucke <jakob.naucke@ibm.com>