k8s: Ensure swaps are off even on non-baremetal environment
On commit 3b688ad776ffb9feb5 it was added an instruction to disable all
swaps when on baremetal as Kubernetes needs of it. However even on VM based
environments swap must be off. So this changed the integration/kubernetes/init.sh
so that swap is disabled when CI equals to "true" instead.

Fixes #1713
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>