ci: Also uninstall containerd non-.io in cleanup
Clean up scripts uninstall `containerd.io` on Ubuntu. This name is used
by download.docker.com, but not by Ubuntu's packaging (which has to be
used on some platforms), which uses `containerd`. This can lead to
Docker failing to start because containerd was not fully removed and
systemd still remembers the custom-installed unit file which points to
`/usr/local`, as seen e.g. in
http://jenkins.katacontainers.io/job/kata-containers-2.0-ubuntu-s390x-PR/83
(although I can also reproduce this on x86?). Also uninstall the
installed Docker variant only.

Fixes: #3705
Signed-off-by: Jakob Naucke <jakob.naucke@ibm.com>