docs: Update readme for metrics
This PR updates the documentation for metrics with the current
implementation for kata 2.x metrics CI and tests. This PR removes
the master branch trending topics as they are not implemented in
kata 2.x metrics CI and also the master tracking as for kata 2.x
all metrics tests are being run.

Fixes #3710

Signed-off-by: Gabriela Cervantes <gabriela.cervantes.tellez@intel.com>