ci: Introduce CI job runners
This implements an extensible way of executing a CI job on the
developer workstation.

With this initial implementation the job can be executed on localhost or
inside a VM. Currently the only VM engine supported is Vagrant.

It is added makefile targets and this should be primary interface with
users.

Fixes #3642
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>