ci: Set GO111MODULE="auto" in lib.sh
The variable GO111MODULE is set to "auto" on .ci/jenkins_job_build.sh so
to address a compability issue raised when it was bumped to Golang 1.16 (see
commit 188c129f3).

That variable should be exported from somewhere else if you are
not a Jenkins job though. This commit will ensure that the variable is set
in .ci/lib.sh that .ci/setup.sh (and its callers) won't need to
export GO111MODULE="auto".

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>