tests/install_kata_kernel: Remove experimental_kernel support
It looks like this was used when virtiofs was new and not in mainline.  The
experimental and vanilla kernel versions now point at exactly the same
thing so there's no point to the extra logic for handling the experimental
case.  If we need an experimental kernel in future, we can point the kernel
version in the versions.yaml file elsewhere instead.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>