tests/install_kata_kernel: Better handling of tools/packaging path
The variable $kernel_repo_dir has a bad name: it suggests its pointing to
a repository of the actual kernel (either source or binary).  In fact it
just points to the tools/packaging subdirectory of the Kata repository.
Furthermore, every use of it appends '/kernel.

Rename it to $kernel_packaging_dir and include the '/kernel' in the
variable.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>