tests/install_kata_kernel: Simplify fallback logic
The script currently checks the kernel version available in the latest
cached build and compares that against the desired version.  It only
attempts to download the cached kernel if the versions match.

But, the URLs for the cached kernels already include the version.  We don't
actually care what the cached version is - only whether the one we want has
a prebuilt image.  So, always attempt to download the cached kernel and
fallback to building from source otherwise.  This simplifies and shortens
the script a bit.

While we're there update some of the informational messages; remove some
that don't really give useful information and update others to better
reflect the new logic.

fixes #3758

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>