lib: Ensure the StartBurstLimit is disabled when restarting services
Too many systemd service restarts in a short period of time
may result in systemd disabling the service.

Add a function to the common library that abstracts the
service restart. Use the opportunity to add error messages
and return error codes so then later will be possible
to stop the test if something goes wrong.

fixes #3794

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>