lib/common: Show the script name when logging
Don't guess the script that output the message.
Showing the script name is only the first step,
all messages should use the info/warn/err functions.

Evaluate the BASH_COMMAND to see the expanded
value while at it.

Try to guess the line number so it will be easier
to map the logs to the code. Not working perfectly,
but it is better than nothing.

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>