kata-webhook: Use http from kubewebhook v2
This change is only done here as a way to unblock the switch to `go mod`
(from the old & deprecated `dep` tool).

Daniel Maizel has opened github.com/kata-containers/tests#3775 which
contains the proper fix for kata-containers, but let's do the switch to
`go mod` first, clear this out, and then get back to the kata-webhook.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>