CCv0 | ci: Add CCv0 features into CI script
Make the build understand if they are running against a target branch
of CCv0 or main and update branches dynamically based on this

Add code to get containerd from our CCv0 PR rather than a released build

Add new CCv0 condition to the code that created the kata image.
CCv0 requires some extra packages for running and debug. It also needs
umoci, which is not yet availabe as a distro package, so wget it direct

Fixes #3887

Signed-off-by: stevenhorsman <steven@uk.ibm.com>