metrics: Fix unknown runtime error
Currently for the memory test we are using containerd and a as runtime io.containerd.kata.v2,
however, we can not extract general information like hypervisor path in order to get the pss
memory for the hypervisor. This PR fixes that issue by using kata-runtime when we are using
io.containerd.kata.v2 so we can get the information properly.

Fixes #3935

Signed-off-by: Gabriela Cervantes <gabriela.cervantes.tellez@intel.com>