tests: Use qemu patching scripts from kata-containers repository
The 'build_and_install_qemu' function in .ci/install_qemu.sh has explicit
code for patching the qemu we're going to use for running the guest.  But,
we already have scripts for this in the main kata repo, specifically
in tools/packaging/scripts/patch_qemu.sh.  Confusingly the latter is used
for static qemu builds, but the one in tests is used for other qemu builds.

Move to using patch_qemu.sh consistently.

fixes #3957

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>