ci: Introduce the vagrant-cleaner.sh script
The vagrant configuration sometimes can get into inconsistent state. For
instance, the domain on libvirt was created by the framework that thinks
the box is not initialized yet. Or eventually the developer wants to
simply wipe out all vagrant files and created resources from his/her
workstation.

In this commit it was introduced the vagrant-cleaner.sh script which
aims on help developers to clean up the vagrant stuffs.

Fixes #3757
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>