ci/setup: install driverctl/pciutils on Fedora/Ubuntu
The commit a043363446c7331 introduced improvements to the vfio
functional test which now requires the driverctl tool. The
driverctl package is installed in the host via .ci/vfio_jenkins_job_build.sh,
but that script is never called by the .ci/setup.sh; as a result, the
package isn't installed when I run the Vagrant-based job locally.

This commit adds the driverctl (also pciutils) to the list of packages
installed in the host Fedora and Ubuntu.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>