agent: Install Rust for shutdown test -- respin
The newly introduced agent shutdown test (#3495) needs Rust, but if we
build the agent itself in a container, Rust will not be installed on the
host. Add a catch in the script. For agent-ctl, run permissions fix as
for cargo setup (this was relevant only on the initrd CI).

This is a respin of #3813 following #3979. The existence check of Cargo
was (erroneously) run _before_ the `source` that would put Cargo into
$PATH.

Fixes: #3812
Signed-off-by: Jakob Naucke <jakob.naucke@ibm.com>