k8s/e2e: call cleanup_env.sh on cleanup()
The environment for running the e2e tests is created from the integration/kubernetes/init.sh
script. It should call the integration/kubernetes/cleanup_env.sh when it is done
otherwise Kubernetes is kept running and resources allocated, this prevents us
from running the e2e tests many times on same host as well as it can mess with
upcoming tests.

Fixes #4037
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>