tracing: Start tmux sessions with sudo
Start all tmux sessions with sudo because vm is not started within a
session as a regular user.

Also use sudo when using ss command to check for VSOCK connections; the
test fails without it because the output does not produce enough
information for the subsequent egrep to eliminate a connection,
resulting in multiple VSOCK entries found.

Fixes #3981

Signed-off-by: Chelsea Mafrica <chelsea.e.mafrica@intel.com>