ci/openshift-ci: allow to use the system QEMU
This changed the kata installer daemonset to allow changing Kata's configuration to
use the QEMU installed in the host.

However, QEMU isn't installed in Red Hat CoreOS (RHCOS) by default. It is needed to
enable an RHCOS extension called 'sandboxed-containers' which will update the cluster's
nodes properly. So this commit also introduces the code to enable that extension as well
as halt while the cluster is not fully upgraded.

To enable the use of system QEMU, it is needed to export "KATA_WITH_SYSTEM_QEMU=yes" before
calling the .ci/openshift-ci/cluster/install_kata.sh script.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>