static-checks: skip check_headers on .dockerignore files
We'll need to add a `.dockerignore` file as part of our kata-deploy
tool.  However, `static-checks.sh` would try to check for the expected
headers in that file, leading to a breakage.

In order to avoid sych a breakage, let's treat the `.dockerignore` file
exactly as we treat the `.gitignore` files, and exclude it from the
header_check.

Fixes: #4095

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>