tests/vfio: Prepare vfio test script for testing more modes
Currently the vfio test case only (minimally) tests the old VFIO handling
behaviour (vfio_mode=guest-kernel).  We're going to want to add tests
for the new behaviour (vfio_mode=vfio).  As a preliminary, rearrange the
test script a little to factor out the parts which will be common from the
parts which are specific to the current test.

Additionally we explicitly add annotations for guest-kernel mode to the
sample container configuration, to prepare for possible future changes to
the default configuration.  We also explicitly annotate to disable vIOMMU
since that shouldn't be necessary for guest-kernel mode.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>