ci: add a way to selectively skip tests based on cri-o version
Some of the tests we skip have been fixed in specific versions of cri-o.
Enabling them for all versions would lead to failures on older branches.

This code makes it possible to mark in which version of cri-o the tests
have been fixed, and enable them only when we're testing in a version
where the fix is present.

Signed-off-by: Julien Ropé <jrope@redhat.com>