tests/vfio: Fix bogus fileMode in sample config.json
The guest-kernel.json.in file has a sample config.json template for a
container with a VFIO device.  It gives the mode of the device as "600"
which seems like a perfectly reasonable octal device mode... but JSON is
not in octal, so it ends up as nonsense.

In fact it doesn't matter for this particular test, since the device gets
interpreted by the guest kernel and doesn't make it into the container when
vfio_mode=guest-kernel, but having this in the config file is still
confusing.  Alter it to the decimal equivalent for clarity.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>