tests/vfio: Improve reporting of failed VFIO guest-kernel test
Currently if something goes wrong with the main test case for VFIO in
guest-kernel mode, all that happens is a fairly unobtrusive grep failure
in the midst of lots of other messages (since the clean up still happens).

Add an explicit error message to make this a bit more obvious.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>