tests/install_kata_kernel: Avoid sudo curl
Running network accessing code with privilege isn't usually a good idea,
especially if we let it access our existing environment.  However, we do
that in two places in install_kata_kernel.sh.

This is done because we're writing files to directories that the regular
user running the script probably won't have write access to.  We can work
around that quite easily, however, by downloading to a temporary location
without privilege, then moving the downloaded file as root.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>