tests: Revert "tests: use containerd to replace cri-containerd"
This reverts commit 393e0714667fe45cf9feece5d3f1ce51f53f82fd.

On the mentioned commit we moved to using containerd instead of
cri-containerd for parsing the crioptions.  However, it turned out to
cause issues when using "not too new enough" versions of containerd,
which happens to be the case on cloud providers services such as AKS.

Tao already proposed to revert the offending commit on the
`kata-containers` repo, but that requires the change reverted on the
tests repo as well, so let's do it.

Fixes: #4173
Depends-on: github.com/kata-containers/kata-containers#3000

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>