tests: Install the oldest, rather than the newest supported Rust by default
If not given an explicit version install_rust.sh looks up versions.yaml to
find the newest Rust version that's known to work.  One of the main uses of
this script is in the CI, where installing the newest Rust version isn't
really what we want:
  - Rust has strong backwards compatibility, so breakage due to a newer
    Rust compiler version is very unlikely
  - However, breakage due to adding things to the Kata codebase which
    require new Rust features that aren't in the *oldests* supported version
    is very likely.

Therefore, it's much more useful to test with the oldest supported version
rather than the newest.  Change install_rust.sh to do this.

fixes #4188

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>